+++ packages/utils/src/getFlatMenus/getFlatMenus.test.ts
@@ -1,4 +1,5 @@
-import transformRoute from './index';
+import getFlatMenus from './getFlatMenus';
+import transformRoute from '../transformRoute/transformRoute';
 
 const routes = [
   {
@@ -19,7 +20,11 @@ const routes = [
         unaccessible: false,
       },
     ],
-
+  },
+  {
+    name: 'list.table.result',
+    path: '/list/:id',
+    exact: true,
     unaccessible: false,
   },
   {
@@ -31,8 +36,9 @@ const routes = [
   { path: '/', redirect: '/welcome', exact: true, unaccessible: false },
 ];
 
-const { menuData, breadcrumb } = transformRoute(routes, false, ({ id }) => {
+const { menuData } = transformRoute(routes, false, ({ id }) => {
   if (id === 'menu.list.table-list') return '查询表格';
+  if (id === 'menu.list.table.result') return '数据详情';
   if (id === 'menu.admin') return '管理页';
   if (id === 'menu.admin.sub-page') return '二级管理页';
   if (id === 'menu.welcome') return '欢迎';
@@ -40,12 +46,8 @@ const { menuData, breadcrumb } = transformRoute(routes, false, ({ id }) => {
 });
 
 test('normal', () => {
-  expect(menuData[0].name).toEqual('欢迎');
-  expect(menuData[1].name).toEqual('管理页');
-  // @ts-ignore
-  expect(menuData[1].children[0].name).toEqual('二级管理页');
+  const flatMenus = getFlatMenus(menuData);
+  expect(Object.keys(flatMenus).length).toEqual(5);
 
-  expect(breadcrumb.get('/admin')?.name).toEqual('管理页');
-  expect(breadcrumb.get('/welcome')?.name).toEqual('欢迎');
-  expect(breadcrumb.get('/admin/sub-page')?.name).toEqual('二级管理页');
+  expect(flatMenus['/list'].name).toEqual('查询表格');
 });
